Add program for updating requirements from external repo
It combines the test and production requirements into a single file
for simplicity.

While these same requirements are used for the data repo, they are not
exactly the same. Adding a separate venv based on the the data repo
requirements doubled the container image size. We can go down that route when
the requirements diverge enough to justify it.